Conflicts:
	cecocloud-main/liquibase/liquibase.properties
	cecocloud-module/cecocloud-module-fact/cecocloud-module-fact-persist/db/schema-postgresql.sql

+++ cecocloud-logic/src/main/java/es/limit/cecocloud/logic/service/SincronitzacioServiceImpl.java
@@ -16,8 +16,10 @@ import es.limit.base.boot.persist.entity.UsuariEntity;
 import es.limit.base.boot.persist.repository.UsuariRepository;
 import es.limit.cecocloud.logic.api.dto.Empresa;
 import es.limit.cecocloud.logic.api.dto.Empresa.EmpresaTipusEnum;
+import es.limit.cecocloud.logic.api.dto.IdentificadorRecursOrigen;
 import es.limit.cecocloud.logic.api.dto.Operari;
 import es.limit.cecocloud.logic.api.dto.OperariEmpresa;
+import es.limit.cecocloud.logic.api.dto.PerfilUsuariIdentificadorEmpresa;
 import es.limit.cecocloud.logic.api.dto.SincronitzacioEmpresa;
 import es.limit.cecocloud.logic.api.dto.SincronitzacioEmpresaOperari;
 import es.limit.cecocloud.logic.api.dto.SincronitzacioEmpresaUsuari;
@@ -34,12 +36,14 @@ import es.limit.cecocloud.persist.entity.EmpresaEntity;
 import es.limit.cecocloud.persist.entity.IdentificadorEntity;
 import es.limit.cecocloud.persist.entity.OperariEmpresaEntity;
 import es.limit.cecocloud.persist.entity.OperariEntity;
+import es.limit.cecocloud.persist.entity.PerfilUsuariIdentificadorEmpresaEntity;
 import es.limit.cecocloud.persist.entity.UsuariIdentificadorEmpresaEntity;
 import es.limit.cecocloud.persist.entity.UsuariIdentificadorEntity;
 import es.limit.cecocloud.persist.repository.EmpresaRepository;
 import es.limit.cecocloud.persist.repository.IdentificadorRepository;
 import es.limit.cecocloud.persist.repository.OperariEmpresaRepository;
 import es.limit.cecocloud.persist.repository.OperariRepository;
+import es.limit.cecocloud.persist.repository.PerfilUsuariIdentificadorEmpresaRepository;
 import es.limit.cecocloud.persist.repository.UsuariIdentificadorEmpresaRepository;
 import es.limit.cecocloud.persist.repository.UsuariIdentificadorRepository;
 
@@ -66,6 +70,8 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 	private OperariRepository operariRepository;
 	@Autowired
 	private OperariEmpresaRepository operariEmpresaRepository;
+	@Autowired
+	private PerfilUsuariIdentificadorEmpresaRepository perfilUsuariIdentificadorEmpresaRepository;
 
 	@Override
 	@Transactional
@@ -122,9 +128,9 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				empresa.getEmbedded().setNom(syncFound.getNom());
 				empresa.getEmbedded().setActiva(true);
 				updateCount++;
-			} else {
-				// Si l'empresa existeix a la BBDD i no a la informació de sincronització
-				// desactiva l'empresa
+			} else if (IdentificadorRecursOrigen.SYNC.equals(empresa.getEmbedded().getOrigen())) {
+				// Si l'empresa existeix a la BBDD i no a la informació de sincronització i el seu orígen
+				// és SYNC -> desactiva l'empresa
 				empresa.getEmbedded().setActiva(false);
 				for (OperariEmpresaEntity operariEmpresa: operariEmpresaRepository.findByEmpresa(empresa)) {
 					operariEmpresa.getEmbedded().setActiu(false);
@@ -148,6 +154,7 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				empresa.setNif(empresaSync.getNif());
 				empresa.setNom(empresaSync.getNom());
 				empresa.setTipus(EmpresaTipusEnum.GESTIO);
+				empresa.setOrigen(IdentificadorRecursOrigen.SYNC);
 				empresa.setActiva(true);
 				empresaRepository.save(
 						EmpresaEntity.builder().
@@ -187,9 +194,9 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				// actualitza la informació de l'operari
 				usuariIdentificador.getEmbedded().setActiu(true);
 				updateCount++;
-			} else {
-				// Si l'operari existeix a la BBDD i no a la informació de sincronització
-				// desactiva l'usuariIdentificador de l'empresa
+			} else if (IdentificadorRecursOrigen.SYNC.equals(usuariIdentificador.getEmbedded().getOrigen())) {
+				// Si l'usuari-identificador existeix a la BBDD i no a la informació de sincronització i el seu orígen
+				// és SYNC -> desactiva l'usuari-identificador de l'empresa
 				usuariIdentificador.getEmbedded().setActiu(false);
 				for (UsuariIdentificadorEmpresaEntity usuariIdentificadorEmpresa: usuariIdentificadorEmpresaRepository.findByUsuariIdentificador(usuariIdentificador)) {
 					usuariIdentificadorEmpresaRepository.delete(usuariIdentificadorEmpresa);
@@ -212,6 +219,7 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				Optional<UsuariEntity> usuari = usuariRepository.findByEmbeddedCodi(usuariSync.getUsuariCodi());
 				if (usuari.isPresent()) {
 					UsuariIdentificador usuariIdentificador = new UsuariIdentificador();
+					usuariIdentificador.setOrigen(IdentificadorRecursOrigen.SYNC);
 					usuariIdentificador.setActiu(true);
 					usuariIdentificadorRepository.save(
 							UsuariIdentificadorEntity.builder().
@@ -253,9 +261,9 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				// actualitza la informació de l'operari
 				operari.getEmbedded().setActiu(true);
 				updateCount++;
-			} else {
-				// Si l'operari existeix a la BBDD i no a la informació de sincronització
-				// desactiva l'operari
+			} else if (IdentificadorRecursOrigen.SYNC.equals(operari.getEmbedded().getOrigen())) {
+				// Si l'operari existeix a la BBDD i no a la informació de sincronització i el seu orígen
+				// és SYNC -> desactiva l'operari de l'empresa
 				operari.getEmbedded().setActiu(false);
 				for (OperariEmpresaEntity operariEmpresa: operariEmpresaRepository.findByOperari(operari)) {
 					operariEmpresa.getEmbedded().setActiu(false);
@@ -278,6 +286,7 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 				if (usuari.isPresent()) {
 					Operari operari = new Operari();
 					operari.setCodi(operariSync.getCodi());
+					operari.setOrigen(IdentificadorRecursOrigen.SYNC);
 					operari.setActiu(true);
 					operariRepository.save(
 							OperariEntity.builder().
@@ -340,12 +349,20 @@ public class SincronitzacioServiceImpl implements SincronitzacioService {
 						if (!usuariIdentificadorEmpresaDb.isPresent()) {
 							// Si l'usuari-identificador-empresa no existeix a la BBDD i si a la informació de sincronització
 							// crea l'usuari-identificador-empresa a la BBDD
-							usuariIdentificadorEmpresaRepository.save(
+							UsuariIdentificadorEmpresaEntity usuariIdentificadorEmpresa = usuariIdentificadorEmpresaRepository.save(
 									UsuariIdentificadorEmpresaEntity.builder().
 									embedded(new UsuariIdentificadorEmpresa()).
 									usuariIdentificador(usuariIdentificadorDb.get()).
 									empresa(empresaDb).
 									build());
+							if (identificador.getPerfilDefecte() != null) {
+								perfilUsuariIdentificadorEmpresaRepository.save(
+										PerfilUsuariIdentificadorEmpresaEntity.builder().
+										embedded(new PerfilUsuariIdentificadorEmpresa()).
+										perfil(identificador.getPerfilDefecte()).
+										usuariIdentificadorEmpresa(usuariIdentificadorEmpresa).
+										build());
+							}
 							createCount++;
 						} else {
 							// Si l'usuari-identificador-empresa existeix a la BBDD i a la informació de sincronització
